+++ odoo/local-src/enfinfidu_payroll/wizard/hr_payroll_config.py
@@ -15,3 +15,15 @@ class HrPayrollConfig(models.TransientModel):
         default=lambda self: self._get_default_configs('laanp_per'),
         digits=dp.get_precision('Payroll Rate'),
         required=False)
+
+    amat_ge_per = fields.Float(
+        string="Maternity insurance (AMAT GE)",
+        default=lambda self: self._get_default_configs('amat_ge_per'),
+        digits=dp.get_precision('Payroll Rate'),
+        required=False)
+
+    def values_to_company(self):
+        super(HrPayrollConfig, self).values_to_company()
+        company_id = self.company_id
+        company_id.write({'laanp_per': self.laanp_per,
+                          'amat_ge_per': self.amat_ge_per})